Made enemies work with new grid changes, made changes for architectural review
