+++ Assets/Concept 2/Objects/Prefabs/Crafter.prefab
@@ -80,7 +80,8 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 2765d751f3d6175469d9122e8ac90e23, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  interceptionCalculator: {fileID: 5564735652332944096}
+  line1: {fileID: 0}
+  line2: {fileID: 0}
   xAnswer: {fileID: 8168581396879448073}
   yAnswer: {fileID: 8923226210279615800}
   onCorrectAnswerGivenEvent:
@@ -1755,7 +1756,6 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 2005012185629278863}
-  - component: {fileID: 5564735652332944096}
   m_Layer: 0
   m_Name: LineSystem
   m_TagString: Untagged
@@ -1780,20 +1780,6 @@ Transform:
   - {fileID: 6931571298773066212}
   m_Father: {fileID: 6392165978576281484}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!114 &5564735652332944096
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 5577388164008093085}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 2dbbdefbebefd544cadf419b51b3a32d, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  line1: {fileID: 6018000140153687598}
-  line2: {fileID: 7876169785252658364}
 --- !u!1 &5674705329855749110
 GameObject:
   m_ObjectHideFlags: 0
