+++ Assets/Concept 3/Objects/Prefabs/Enemy1.prefab
@@ -14,6 +14,7 @@ GameObject:
   - component: {fileID: 1747232232689400912}
   - component: {fileID: 2381157406376694979}
   - component: {fileID: 8273156906793085097}
+  - component: {fileID: -1385989702630780348}
   m_Layer: 0
   m_Name: Enemy1
   m_TagString: Untagged
@@ -153,7 +154,29 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 9908131592cb08743a4f42cf17c03780, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  pathData: {fileID: 0}
+  gridData: {fileID: 0}
   movementSpeed: 2
   damage: 20
   attackSpeed: 1
+--- !u!195 &-1385989702630780348
+NavMeshAgent:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 2038675575060427630}
+  m_Enabled: 1
+  m_AgentTypeID: 0
+  m_Radius: 0.5
+  m_Speed: 3.5
+  m_Acceleration: 8
+  avoidancePriority: 50
+  m_AngularSpeed: 120
+  m_StoppingDistance: 0
+  m_AutoTraverseOffMeshLink: 1
+  m_AutoBraking: 1
+  m_AutoRepath: 1
+  m_Height: 1
+  m_BaseOffset: 0.5
+  m_WalkableMask: 4294967295
+  m_ObstacleAvoidanceType: 4
