+++ Assets/Concept 3/Objects/Prefabs/LevelHolder.prefab
@@ -44,7 +44,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: ac4c39d4185aa9a47b4720b2bc1c4f1f, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  gridTile: {fileID: 1673266920155040351, guid: ef1b0feb1f606214294ffb81531b6f6d, type: 3}
+  gridTile: {fileID: 3530273512751714376, guid: ef1b0feb1f606214294ffb81531b6f6d, type: 3}
   placeableGridTile: {fileID: 6663351834643575591, guid: ac216fbc6dc79c1478fff6fa9cc0843e, type: 3}
   pathGridTile: {fileID: 2990015622576239724, guid: f32c187719ca72b4c807b4c710a25d80, type: 3}
-  pathData: {fileID: 0}
+  gridData: {fileID: 0}
