+++ Assets/Concept 3/Objects/Variables/GridDataLevel1.asset
@@ -9,10 +9,10 @@ MonoBehaviour:
   m_GameObject: {fileID: 0}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 0adf66fd83f57f54db8a00dcb3dfe629, type: 3}
+  m_Script: {fileID: 11500000, guid: 26f63a5f755c2ee46b99f0b62346a2e4, type: 3}
   m_Name: GridDataLevel1
   m_EditorClassIdentifier: 
-  pathCoordinates:
+  pathIndices:
   - {x: 2, y: 0}
   - {x: 2, y: 1}
   - {x: 2, y: 2}
