+++ Assets/Concept 3/Objects/Variables/GridDataLevel2.asset
@@ -9,10 +9,10 @@ MonoBehaviour:
   m_GameObject: {fileID: 0}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 0adf66fd83f57f54db8a00dcb3dfe629, type: 3}
+  m_Script: {fileID: 11500000, guid: 26f63a5f755c2ee46b99f0b62346a2e4, type: 3}
   m_Name: GridDataLevel2
   m_EditorClassIdentifier: 
-  pathCoordinates:
+  pathIndices:
   - {x: 0, y: 9}
   - {x: 1, y: 9}
   - {x: 2, y: 9}
@@ -48,6 +48,6 @@ MonoBehaviour:
   - {x: 17, y: 8}
   - {x: 17, y: 9}
   - {x: 18, y: 9}
-  - {x: 19, y: 9}
+  - {x: 29, y: 9}
   xGridSize: 20
   yGridSize: 20
