--- Assets/Concept 3/Objects/Variables/PathDataLevel1.asset
@@ -1,56 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 0adf66fd83f57f54db8a00dcb3dfe629, type: 3}
-  m_Name: PathDataLevel1
-  m_EditorClassIdentifier: 
-  pathCoordinates:
-  - {x: 2, y: 0}
-  - {x: 2, y: 1}
-  - {x: 2, y: 2}
-  - {x: 2, y: 3}
-  - {x: 2, y: 4}
-  - {x: 2, y: 5}
-  - {x: 2, y: 6}
-  - {x: 2, y: 7}
-  - {x: 2, y: 8}
-  - {x: 2, y: 9}
-  - {x: 2, y: 10}
-  - {x: 2, y: 11}
-  - {x: 2, y: 12}
-  - {x: 2, y: 13}
-  - {x: 2, y: 14}
-  - {x: 3, y: 14}
-  - {x: 4, y: 14}
-  - {x: 5, y: 14}
-  - {x: 6, y: 14}
-  - {x: 7, y: 14}
-  - {x: 7, y: 13}
-  - {x: 7, y: 12}
-  - {x: 7, y: 11}
-  - {x: 7, y: 10}
-  - {x: 7, y: 9}
-  - {x: 7, y: 8}
-  - {x: 7, y: 7}
-  - {x: 7, y: 6}
-  - {x: 7, y: 5}
-  - {x: 8, y: 5}
-  - {x: 9, y: 5}
-  - {x: 10, y: 5}
-  - {x: 11, y: 5}
-  - {x: 12, y: 5}
-  - {x: 13, y: 5}
-  - {x: 14, y: 5}
-  - {x: 15, y: 5}
-  - {x: 16, y: 5}
-  - {x: 17, y: 5}
-  - {x: 18, y: 5}
-  - {x: 19, y: 5}
