--- Assets/Concept 3/Objects/Variables/PathDataLevel1.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: d836a2889ec4107418d00a4bdf0d0452
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
