--- Assets/Concept 3/Objects/Variables/PathDataLevel2.asset
@@ -1,51 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 0adf66fd83f57f54db8a00dcb3dfe629, type: 3}
-  m_Name: PathDataLevel2
-  m_EditorClassIdentifier: 
-  pathCoordinates:
-  - {x: 0, y: 9}
-  - {x: 1, y: 9}
-  - {x: 2, y: 9}
-  - {x: 2, y: 10}
-  - {x: 2, y: 11}
-  - {x: 2, y: 12}
-  - {x: 2, y: 13}
-  - {x: 3, y: 13}
-  - {x: 4, y: 13}
-  - {x: 5, y: 13}
-  - {x: 6, y: 13}
-  - {x: 7, y: 13}
-  - {x: 7, y: 12}
-  - {x: 7, y: 11}
-  - {x: 7, y: 10}
-  - {x: 7, y: 9}
-  - {x: 8, y: 9}
-  - {x: 9, y: 9}
-  - {x: 10, y: 9}
-  - {x: 11, y: 9}
-  - {x: 12, y: 9}
-  - {x: 12, y: 8}
-  - {x: 12, y: 7}
-  - {x: 12, y: 6}
-  - {x: 12, y: 5}
-  - {x: 13, y: 5}
-  - {x: 14, y: 5}
-  - {x: 15, y: 5}
-  - {x: 16, y: 5}
-  - {x: 17, y: 5}
-  - {x: 17, y: 6}
-  - {x: 17, y: 7}
-  - {x: 17, y: 8}
-  - {x: 17, y: 9}
-  - {x: 18, y: 9}
-  - {x: 19, y: 9}
