--- Assets/Concept 3/Objects/Variables/PathDataLevel2.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 45a2ad084e612414ea186eb029b8ab38
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
