--- Assets/Scripts/Crafter/InterceptionCalculator.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 2dbbdefbebefd544cadf419b51b3a32d
