--- Assets/Scripts/WorldGrid/PathData.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 0adf66fd83f57f54db8a00dcb3dfe629
