+++ Packages/manifest.json
@@ -4,6 +4,7 @@
     "com.unity.2d.aseprite": "1.1.4",
     "com.unity.2d.psdimporter": "9.0.3",
     "com.unity.2d.spriteshape": "10.0.6",
+    "com.unity.ai.navigation": "2.0.0",
     "com.unity.burst": "1.8.15",
     "com.unity.cinemachine": "2.9.7",
     "com.unity.collections": "2.1.4",
